Divide classes into packages
Classes all stored into one folder.

Packaging related classes into packages, helps to prevent naming
conflicts, and makes searching for classes easier and cleaner.
