--- src/main/java/Command.java
@@ -1,13 +1,0 @@
-public abstract class Command {
-  
-    public String command;
-    
-    public Command(String command) {
-        this.command = command;    
-    }
-    
-    public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
-   
-    public abstract boolean isExit();
-    
-}
