--- src/main/java/Deadline.java
@@ -1,29 +1,0 @@
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-
-public class Deadline extends Task {
-
-    protected LocalDate by;
-
-    public Deadline(String description, LocalDate by) {
-        super(description);
-        this.by = by;
-    }
-    
-    @Override
-    public String getTaskType() {
-        return "D";
-    }
-    
-    @Override
-    public LocalDate getTiming() {
-        return this.by;
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " 
-                + by.format(DateTimeFormatter.ofPattern("MMM dd yyyy")) + ")";
-    }
-
-}
