--- src/main/java/DukeException.java
@@ -1,7 +1,0 @@
-public class DukeException extends Exception {
-    
-    public DukeException(String message) {
-        super(message);
-    }
-    
-}
