+++ src/main/java/duke/command/ByeCommand.java
@@ -1,0 +1,16 @@
+public class ByeCommand extends Command {
+
+    public ByeCommand() {
+        super("bye");
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        return;        
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+}
