+++ src/main/java/duke/exception/DukeException.java
@@ -1,0 +1,7 @@
+public class DukeException extends Exception {
+    
+    public DukeException(String message) {
+        super(message);
+    }
+    
+}
